+++ frontend_react/react_src/src/components/chat/SuggestedPrompt.tsx
@@ -1,6 +1,7 @@
 import React from "react";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faPaperPlane } from "@fortawesome/free-solid-svg-icons/faPaperPlane";
+import { useGeneratedDictionaries } from "@/api-state/dictionaries/hooks";
 import { usePostMessage } from "@/api-state/chat-messages/hooks";
 import { useAppState } from "@/state/hooks";
 
@@ -9,8 +10,14 @@ interface SuggestedPromptProps {
   chatId?: string;
 }
 
-export const SuggestedPrompt: React.FC<SuggestedPromptProps> = ({ message, chatId }) => {
-  const { enableChartGeneration, enableBusinessInsights, dataSource } = useAppState();
+export const SuggestedPrompt: React.FC<SuggestedPromptProps> = ({
+  message,
+  chatId,
+}) => {
+  const { enableChartGeneration, enableBusinessInsights, dataSource } =
+    useAppState();
+  const { data: dictionaries } = useGeneratedDictionaries();
+  const isDisabled = !dictionaries?.[0];
   const { mutate } = usePostMessage();
   return (
     <div className="h-16 p-3 bg-[#22272b] rounded border justify-start items-center gap-2 inline-flex">
@@ -20,12 +27,20 @@ export const SuggestedPrompt: React.FC<SuggestedPromptProps> = ({ message, chatI
       <div className="w-9 h-9 p-2 justify-center items-center flex">
         <div className="w-5 h-5 flex-col justify-center items-center gap-2.5 inline-flex">
           <div className="text-center text-sm leading-tight cursor-pointer">
-            <FontAwesomeIcon
-              icon={faPaperPlane}
-              onClick={() => {
-                mutate({ message, chatId, enableChartGeneration, enableBusinessInsights, dataSource });
-              }}
-            />
+            {!isDisabled && (
+              <FontAwesomeIcon
+                icon={faPaperPlane}
+                onClick={() => {
+                  mutate({
+                    message,
+                    chatId,
+                    enableChartGeneration,
+                    enableBusinessInsights,
+                    dataSource,
+                  });
+                }}
+              />
+            )}
           </div>
         </div>
       </div>