+++ frontend_react/react_src/src/components/ui-custom/file-uploader.tsx
@@ -14,7 +14,7 @@ interface FileUploaderProps {
 
 export const FileUploader: React.FC<FileUploaderProps> = ({
   maxSize = 1024 * 1024 * 200,
-  accept = { "file/csv": [".csv"] },
+  accept = { "file/csv": [".csv"], "file/xlsx": [".xlsx", ".xls"] },
   progress = 0,
   onFilesChange,
 }) => {